+++ skyfield/timelib.py
@@ -197,6 +197,45 @@ class Timescale(object):
         t.tdb = tdb
         return t
 
+    def ut1(self, year=None, month=1, day=1, hour=0, minute=0, second=0.0,
+            jd=None):
+        """Return the Time corresponding to a specific moment in UT1.
+
+        You can supply the Universal Time (UT1) by providing either a
+        proleptic Gregorian calendar date or a raw Julian Date float.
+        The following two method calls are equivalent::
+
+            timescale.ut1(2014, 1, 18, 1, 35, 37.5)
+            timescale.ut1(jd=2456675.56640625)
+
+        """
+        if jd is not None:
+            ut1 = jd
+        else:
+            ut1 = julian_date(
+                _to_array(year), _to_array(month), _to_array(day),
+                _to_array(hour), _to_array(minute), _to_array(second),
+            )
+        ut1 = _to_array(ut1)
+
+        # Estimate TT = UT1, to get a rough Delta T estimate.
+        tt_approx = ut1
+        delta_t_approx = interpolate_delta_t(self.delta_t_table, tt_approx)
+
+        # Use the rough Delta T to make a much better estimate of TT,
+        # then generate an even better Delta T.
+        tt_approx = ut1 + delta_t_approx / DAY_S
+        delta_t_approx = interpolate_delta_t(self.delta_t_table, tt_approx)
+
+        # We can now estimate TT with an error of < 1e-9 seconds within
+        # 10 centuries of either side of the present; for details, see:
+        # https://github.com/skyfielders/astronomy-notebooks
+        # and look for the notebook "error-in-timescale-ut1.ipynb".
+        tt = ut1 + delta_t_approx / DAY_S
+        t = Time(self, tt)
+        t.ut1 = ut1
+        return t
+
     def from_astropy(self, t):
         """Return a Skyfield time corresponding to the AstroPy time `t`."""
         return self.tt(jd=t.tt.jd)
@@ -427,6 +466,25 @@ class Time(object):
         else:
             return strftime(format, tup)
 
+    def _utc_year(self):
+        """Return a fractional UTC year, for convenience when plotting.
+
+        An experiment, probably superseded by the ``J`` attribute below.
+
+        """
+        d = self._utc_float() - 1721059.5
+        #d += offset
+        C = 365 * 100 + 24
+        d -= 365
+        d += d // C - d // (4 * C)
+        d += 365
+        # Y = d / C * 100
+        # print(Y)
+        K = 365 * 3 + 366
+        d -= (d + K*7//8) // K
+        # d -= d // 1461.0
+        return d / 365.0 # d // 365, d % 365.0
+
     def _utc_tuple(self, offset=0.0):
         """Return UTC as (year, month, day, hour, minute, second.fraction).
 
@@ -506,6 +564,11 @@ class Time(object):
     # Conversion between timescales.
 
     @reify
+    def J(self):
+        """Decimal Julian years centered on J2000.0 = TT 2000 January 1 12h."""
+        return (self.tt - 1721045.0) / 365.25
+
+    @reify
     def tai(self):
         self.tai = tai = self.tt - tt_minus_tai
         return tai
@@ -532,6 +595,10 @@ class Time(object):
         table = self.ts.delta_t_table
         self.delta_t = delta_t = interpolate_delta_t(table, self.tt)
         return delta_t
+
+    @reify
+    def dut1(self):
+        return (self.tt - self._utc_float()) * DAY_S - self.delta_t
 
     @reify
     def gmst(self):